requirements.txt
# Code style tools dependency
pre_commit==4.3.0
isort==7.0.0
flake8==7.3.0

# Django dependency
Django==5.2.7
django-environ==0.12.0
django-filter==25.2

# DRF dependency
djangorestframework==3.16.1
django-cors-headers==4.9.0

# Third-party dependency
requests==2.32.5
gunicorn==23.0.0
pillow==12.0.0
